+++ src/account.cpp
@@ -828,17 +828,17 @@ bool Account::isTlsEnabled() const
    return protocol() == Account::Protocol::RING || (d_ptr->accountDetail(DRing::Account::ConfProperties::TLS::ENABLED) IS_TRUE);
 }
 
-///Return if the ringtone are enabled
-bool Account::isRingtoneEnabled() const
-{
-   return (d_ptr->accountDetail(DRing::Account::ConfProperties::Ringtone::ENABLED) IS_TRUE);
-}
+/////Return if the ringtone are enabled
+//bool Account::isRingtoneEnabled() const
+//{
+//   return (d_ptr->accountDetail(DRing::Account::ConfProperties::Ringtone::ENABLED) IS_TRUE);
+//}
 
-///Return the account ringtone path
-QString Account::ringtonePath() const
-{
-   return d_ptr->accountDetail(DRing::Account::ConfProperties::Ringtone::PATH);
-}
+/////Return the account ringtone path
+//QString Account::ringtonePath() const
+//{
+//   return d_ptr->accountDetail(DRing::Account::ConfProperties::Ringtone::PATH);
+//}
 
 ///Return the last error message received
 QString Account::lastErrorMessage() const
@@ -944,30 +944,30 @@ bool Account::presenceEnabled() const
    return d_ptr->accountDetail(DRing::Account::ConfProperties::Presence::ENABLED) IS_TRUE;
 }
 
-bool Account::isVideoEnabled() const
-{
-   return d_ptr->accountDetail(DRing::Account::ConfProperties::Video::ENABLED) IS_TRUE;
-}
+//bool Account::isVideoEnabled() const
+//{
+//   return d_ptr->accountDetail(DRing::Account::ConfProperties::Video::ENABLED) IS_TRUE;
+//}
 
-int Account::videoPortMax() const
-{
-   return d_ptr->accountDetail(DRing::Account::ConfProperties::Video::PORT_MAX).toInt();
-}
+//int Account::videoPortMax() const
+//{
+//   return d_ptr->accountDetail(DRing::Account::ConfProperties::Video::PORT_MAX).toInt();
+//}
 
-int Account::videoPortMin() const
-{
-   return d_ptr->accountDetail(DRing::Account::ConfProperties::Video::PORT_MIN).toInt();
-}
+//int Account::videoPortMin() const
+//{
+//   return d_ptr->accountDetail(DRing::Account::ConfProperties::Video::PORT_MIN).toInt();
+//}
 
-int Account::audioPortMin() const
-{
-   return d_ptr->accountDetail(DRing::Account::ConfProperties::Audio::PORT_MIN).toInt();
-}
+//int Account::audioPortMin() const
+//{
+//   return d_ptr->accountDetail(DRing::Account::ConfProperties::Audio::PORT_MIN).toInt();
+//}
 
-int Account::audioPortMax() const
-{
-   return d_ptr->accountDetail(DRing::Account::ConfProperties::Audio::PORT_MAX).toInt();
-}
+//int Account::audioPortMax() const
+//{
+//   return d_ptr->accountDetail(DRing::Account::ConfProperties::Audio::PORT_MAX).toInt();
+//}
 
 bool Account::isUpnpEnabled() const
 {
@@ -1116,8 +1116,8 @@ QVariant Account::roleData(int role) const
          return sipStunServer();
       case CAST(Account::Role::PublishedAddress):
          return publishedAddress();
-      case CAST(Account::Role::RingtonePath):
-         return ringtonePath();
+      //case CAST(Account::Role::RingtonePath):
+      //   return ringtonePath();
       case CAST(Account::Role::RegistrationExpire):
          return registrationExpire();
       case CAST(Account::Role::TlsNegotiationTimeoutSec):
@@ -1146,8 +1146,8 @@ QVariant Account::roleData(int role) const
          return isSipStunEnabled();
       case CAST(Account::Role::PublishedSameAsLocal):
          return isPublishedSameAsLocal();
-      case CAST(Account::Role::RingtoneEnabled):
-         return isRingtoneEnabled();
+      //case CAST(Account::Role::RingtoneEnabled):
+      //   return isRingtoneEnabled();
       case CAST(Account::Role::dTMFType):
          return DTMFType();
       case CAST(Account::Role::Id):
@@ -1187,16 +1187,16 @@ QVariant Account::roleData(int role) const
          return supportPresenceSubscribe();
       case CAST(Account::Role::PresenceEnabled          ):
          return presenceEnabled();
-      case CAST(Account::Role::IsVideoEnabled           ):
-         return isVideoEnabled();
-      case CAST(Account::Role::VideoPortMax             ):
-         return videoPortMax();
-      case CAST(Account::Role::VideoPortMin             ):
-         return videoPortMin();
-      case CAST(Account::Role::AudioPortMin             ):
-         return audioPortMin();
-      case CAST(Account::Role::AudioPortMax             ):
-         return audioPortMax();
+      //case CAST(Account::Role::IsVideoEnabled           ):
+      //   return isVideoEnabled();
+      //case CAST(Account::Role::VideoPortMax             ):
+      //   return videoPortMax();
+      //case CAST(Account::Role::VideoPortMin             ):
+      //   return videoPortMin();
+      //case CAST(Account::Role::AudioPortMin             ):
+      //   return audioPortMin();
+      //case CAST(Account::Role::AudioPortMax             ):
+      //   return audioPortMax();
       case CAST(Account::Role::IsUpnpEnabled            ):
          return isUpnpEnabled();
       case CAST(Account::Role::HasCustomUserAgent       ):
@@ -1989,9 +1989,9 @@ void Account::setRoleData(int role, const QVariant& value)
       case CAST(Account::Role::PublishedAddress):
          setPublishedAddress(value.toString());
          break;
-      case CAST(Account::Role::RingtonePath):
-         setRingtonePath(value.toString());
-         break;
+      //case CAST(Account::Role::RingtonePath):
+      //   setRingtonePath(value.toString());
+      //   break;
       case CAST(Account::Role::RegistrationExpire):
          setRegistrationExpire(value.toInt());
          break;
@@ -2034,9 +2034,9 @@ void Account::setRoleData(int role, const QVariant& value)
       case CAST(Account::Role::PublishedSameAsLocal):
          setPublishedSameAsLocal(value.toBool());
          break;
-      case CAST(Account::Role::RingtoneEnabled):
-         setRingtoneEnabled(value.toBool());
-         break;
+      //case CAST(Account::Role::RingtoneEnabled):
+      //   setRingtoneEnabled(value.toBool());
+      //   break;
       case CAST(Account::Role::dTMFType):
          setDTMFType((DtmfType)value.toInt());
          break;
@@ -2056,21 +2056,21 @@ void Account::setRoleData(int role, const QVariant& value)
       case CAST(Account::Role::PresenceEnabled          ):
          setPresenceEnabled(value.toBool());
          break;
-      case CAST(Account::Role::IsVideoEnabled           ):
-         setVideoEnabled(value.toBool());
-         break;
-      case CAST(Account::Role::VideoPortMax             ):
-         setVideoPortMax(value.toInt());
-         break;
-      case CAST(Account::Role::VideoPortMin             ):
-         setVideoPortMin(value.toInt());
-         break;
-      case CAST(Account::Role::AudioPortMin             ):
-         setAudioPortMin(value.toInt());
-         break;
-      case CAST(Account::Role::AudioPortMax             ):
-         setAudioPortMax(value.toInt());
-         break;
+      //case CAST(Account::Role::IsVideoEnabled           ):
+      //   setVideoEnabled(value.toBool());
+      //   break;
+      //case CAST(Account::Role::VideoPortMax             ):
+      //   setVideoPortMax(value.toInt());
+      //   break;
+      //case CAST(Account::Role::VideoPortMin             ):
+      //   setVideoPortMin(value.toInt());
+      //   break;
+      //case CAST(Account::Role::AudioPortMin             ):
+      //   setAudioPortMin(value.toInt());
+      //   break;
+      //case CAST(Account::Role::AudioPortMax             ):
+      //   setAudioPortMax(value.toInt());
+      //   break;
       case CAST(Account::Role::IsUpnpEnabled            ):
          setUpnpEnabled(value.toBool());
          break;
@@ -2598,3 +2598,198 @@ bool AccountPrivate::merge(Account* account)
 #undef TO_BOOL
 #undef IS_TRUE
 #include <account.moc>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+/*******************************************************************************
+*                                                                             *
+*                                  Getters                                    *
+*                                                                             *
+******************************************************************************/
+
+bool AccountMedia::isVideoEnabled() const
+{
+	return d_ptr->accountDetail(DRing::Account::ConfProperties::Video::ENABLED) IS_TRUE;
+}
+
+int AccountMedia::videoPortMax() const
+{
+	return d_ptr->accountDetail(DRing::Account::ConfProperties::Video::PORT_MAX).toInt();
+}
+
+int AccountMedia::videoPortMin() const
+{
+	return d_ptr->accountDetail(DRing::Account::ConfProperties::Video::PORT_MIN).toInt();
+}
+
+int AccountMedia::audioPortMin() const
+{
+	return d_ptr->accountDetail(DRing::Account::ConfProperties::Audio::PORT_MIN).toInt();
+}
+
+int AccountMedia::audioPortMax() const
+{
+	return d_ptr->accountDetail(DRing::Account::ConfProperties::Audio::PORT_MAX).toInt();
+}
+
+///Return if the ringtone are enabled
+bool AccountMedia::isRingtoneEnabled() const
+{
+	return (d_ptr->accountDetail(DRing::Account::ConfProperties::Ringtone::ENABLED) IS_TRUE);
+}
+
+///Return the account ringtone path
+QString AccountMedia::ringtonePath() const
+{
+	return d_ptr->accountDetail(DRing::Account::ConfProperties::Ringtone::PATH);
+}
+
+
+
+#define CAST(item) static_cast<int>(item)
+QVariant AccountMedia::roleData(int role) const
+{
+	switch (role) {
+	case CAST(AccountMedia::Role::IsVideoEnabled):
+		return isVideoEnabled();
+	case CAST(AccountMedia::Role::VideoPortMax):
+		return videoPortMax();
+	case CAST(AccountMedia::Role::VideoPortMin):
+		return videoPortMin();
+	case CAST(AccountMedia::Role::AudioPortMin):
+		return audioPortMin();
+	case CAST(AccountMedia::Role::AudioPortMax):
+		return audioPortMax();
+	case CAST(AccountMedia::Role::RingtonePath):
+		return ringtonePath();
+	case CAST(AccountMedia::Role::RingtoneEnabled):
+		return isRingtoneEnabled();
+	default:
+		return QVariant();
+	}
+	return QVariant();
+}
+#undef CAST
+
+
+
+
+
+
+/*****************************************************************************
+*                                                                           *
+*                                  Setters                                  *
+*                                                                           *
+****************************************************************************/
+
+///Use video by default when available
+void AccountMedia::setVideoEnabled(bool enable)
+{
+	d_ptr->setAccountProperty(DRing::Account::ConfProperties::Video::ENABLED, (enable)TO_BOOL);
+}
+
+/**Set the maximum audio port
+* This can be used when some routers without UPnP support open a narrow range
+* of ports to allow the stream to go through.
+*/
+void AccountMedia::setAudioPortMax(int port)
+{
+	d_ptr->setAccountProperty(DRing::Account::ConfProperties::Audio::PORT_MAX, QString::number(port));
+}
+
+/**Set the minimum audio port
+* This can be used when some routers without UPnP support open a narrow range
+* of ports to allow the stream to go through.
+*/
+void AccountMedia::setAudioPortMin(int port)
+{
+	d_ptr->setAccountProperty(DRing::Account::ConfProperties::Audio::PORT_MIN, QString::number(port));
+}
+
+/**Set the maximum video port
+* This can be used when some routers without UPnP support open a narrow range
+* of ports to allow the stream to go through.
+*/
+void AccountMedia::setVideoPortMax(int port)
+{
+	d_ptr->setAccountProperty(DRing::Account::ConfProperties::Video::PORT_MAX, QString::number(port));
+}
+
+/**Set the minimum video port
+* This can be used when some routers without UPnP support open a narrow range
+* of ports to allow the stream to go through.
+*/
+void AccountMedia::setVideoPortMin(int port)
+{
+	d_ptr->setAccountProperty(DRing::Account::ConfProperties::Video::PORT_MIN, QString::number(port));
+}
+
+///Set the ringtone path, it have to be a valid absolute path
+void AccountMedia::setRingtonePath(const QString& detail)
+{
+	d_ptr->setAccountProperty(DRing::Account::ConfProperties::Ringtone::PATH, detail);
+}
+
+///Set if custom ringtone are enabled
+void AccountMedia::setRingtoneEnabled(bool detail)
+{
+	d_ptr->setAccountProperty(DRing::Account::ConfProperties::Ringtone::ENABLED, (detail)TO_BOOL);
+}
+
+#define CAST(item) static_cast<int>(item)
+void AccountMedia::setRoleData(int role, const QVariant& value)
+{
+	switch (role) {
+	case CAST(AccountMedia::Role::IsVideoEnabled):
+		setVideoEnabled(value.toBool());
+		break;
+	case CAST(AccountMedia::Role::VideoPortMax):
+		setVideoPortMax(value.toInt());
+		break;
+	case CAST(AccountMedia::Role::VideoPortMin):
+		setVideoPortMin(value.toInt());
+		break;
+	case CAST(AccountMedia::Role::AudioPortMin):
+		setAudioPortMin(value.toInt());
+		break;
+	case CAST(AccountMedia::Role::AudioPortMax):
+		setAudioPortMax(value.toInt());
+		break;
+	case CAST(AccountMedia::Role::RingtonePath):
+		setRingtonePath(value.toString());
+		break;
+	case CAST(AccountMedia::Role::RingtoneEnabled):
+		setRingtoneEnabled(value.toBool());
+		break;
+	}
+}
+#undef CAST
